Create setReadable_GENENAME.R
This is just a modified version of DOSE::setReadable() function which hardcodes SYMBOL as the key to translate gene IDs into. The "org.Sc.sgd.db" annotationDBi object has no SYMBOL column so it won't work. It has GENENAME instead. This is a quickfix to no have to change other scripts. 

Load it with source("setReadable_GENENAME.R") in R.